refactor: move report logic to endpoints
