+++ src/API/API.php
@@ -22,20 +22,21 @@ class API {
         // To prevent conflict on we are loading autoload.php when need for now. In future we can loaded it globally.
         require GIVE_PLUGIN_DIR . 'vendor/autoload.php';
 
-        // Register Reports Routes
-        $reports = new Endpoints\Reports();
-        $reports->init();
+        // Load endpoints
+        $this->load_endpoints();
+
     }
 
     public function __construct() {
         //Do nothing
-    }
+	}
 
-    public function setup_endpoints() {
-        // Register Reports Routes
-        $reports = new Endpoints\Reports();
-		$reports->init();
-    }
+	public function load_endpoints() {
+		// Load payment statuses endpoint
+		$paymentStatuses = new Endpoints\Reports\PaymentStatuses();
+		$paymentStatuses->init();
+	}
+
 }
 $api = new API;
 $api->init();
